+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 9.4.0 -
 ### Added
 - Added the ability to list an achievement as the method of acquisition on the secondary line. Mostly useful for mounts and pets where the secondary button is already used. Necessary data is only on TWW and DF mounts, pets and toys in Collections right now, but will get added elsewhere in the future.
+### Changed
+- Remove LibBabble-ItemSet and replace with C_Item.GetItemSetInfo
 ### Fixed
 - Fixed Guardian Vorquin getting overwritten by Otterworldly Ottuk Carrier in Dragonflight mounts in Collections